requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
geoalchemy2>=0.14.0
alembic>=1.12.0
psycopg2-binary>=2.9.7
pydantic>=2.3.0
python-dotenv>=1.0.0
geojson-pydantic>=0.6.0
shapely>=2.0.1
python-jose>=3.3.0
passlib>=1.7.4
pytest>=7.4.2
httpx>=0.25.0
pydantic-settings>=2.0.0
black>=25.1.0
isort>=6.0.1
